+++ src/components/signUp.js
@@ -1,184 +1,259 @@
-import React, { Component } from "react";
-import { withRouter } from "react-router-dom";
+// import React, { Component } from "react";
+// import { withRouter } from "react-router-dom";
 
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
-import { registerUser } from "../action/auth";
-import classnames from "classnames";
+// import PropTypes from "prop-types";
+// import { connect } from "react-redux";
+// import { registerUser } from "../action/auth";
+// import classnames from "classnames";
 
-import Logo from "./img/logo.png";
-import "../App.css";
+// import Logo from "./img/logo.png";
+// import "../App.css";
 
-class Register extends Component {
-  constructor() {
-    super();
-    this.state = {
-      name: "",
-      email: "",
-      phone: "",
-      password: "",
-      passwordConfirm: "",
-      errors: {},
-    };
-  }
-  componentDidMount() {
-    // If logged in and user navigates to Register page, should redirect them to
-    // dashboard
-    if (this.props.auth.isAuthenticated) {
-      this.props.history.push("/");
-    }
-  }
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.errors) {
-      this.setState({ errors: nextProps.errors });
-    }
-  }
-  onChange = (e) => {
-    this.setState({
-      [e.target.id]: e.target.value,
-    });
-  };
-  onChangePhone = (e) => {
-    this.setState({
-      [e.target.id]: e.target.value.replace(/\D/,''),
-    });
-  };
-  onSubmit = (e) => {
-    // e.preventDefault();
-    const newUser = {
-      name: this.state.name,
-      email: this.state.email,
-      phone: this.state.phone,
-      password: this.state.password,
-      passwordConfirm: this.state.passwordConfirm,
-    };
-    this.props.registerUser(newUser, this.props.history);
-    // console.log(newUser);
-  };
-  render() {
-    const { errors } = this.state;
-    return (
-      <div className="App-login">
-        <main>
-          <center>
-            <div className="container">
-              <div
-                className="z-depth-1  lighten-4"
-                style={{
-                  display: "inline-block",
-                  padding: "32px 48px 0px 48px",
-                  border: "1px solid #EEE",
-                }}
-              >
-                <img
-                  src={Logo}
-                  className="responsive-img"
-                  style={{
-                    width: "150px",
-                    margin: 20,
-                  }}
-                  alt="..."
-                />
-                <form onSubmit={this.onSubmit}>
-                  <div className="form-group text-left">
-                    <label>Full Name</label>
-                    <input
-                      className="form-control"
-                      onChange={this.onChange}
-                      value={this.state.name}
-                      error={errors.name}
-                      placeholder=" Full Name"
-                      id="name"
-                      size="25"
-                      type="name"
-                    />
-                  </div>
-                  <div className="form-group text-left">
-                  <label>Email Address</label>
-                    <input
-                      onChange={this.onChange}
-                      value={this.state.email}
-                      error={errors.email}
-                      placeholder=" Email Address"
-                      id="email"
-                      size="25"
-                      type="email"
-                      className={classnames("form-control", {
-                        invalid: errors.email || errors.emailnotfound,
-                      })}
-                    />
-                  </div>
-                  <div className="form-group text-left">
-                  <label>Phone Number</label>
-                    <input
-                      onChange={this.onChangePhone}
-                      value={this.state.phone}
-                      error={errors.phone}
-                      placeholder=" Phone Number"
-                      id="phone"
-                      size="25"
-                      type="text"
-                      // pattern="[0-9]"
-                      className={classnames("form-control", {})}
-                    />
-                  </div>
-                  <div className="form-group text-left">
-                  <label>Password</label>
-                    <input
-                      onChange={this.onChange}
-                      value={this.state.password}
-                      error={errors.password}
-                      placeholder=" Password"
-                      id="password"
-                      size="25"
-                      type="password"
-                      className={classnames("form-control", {
-                        invalid: errors.password || errors.passwordincorrect,
-                      })}
-                    />
-                  </div>
-                  <div className="form-group text-left">
-                  <label>Confirm Password</label>
-                    <input
-                      onChange={this.onChange}
-                      value={this.state.passwordConfirm}
-                      error={errors.passwordConfirm}
-                      placeholder=" Confirm Password"
-                      id="passwordConfirm"
-                      size="25"
-                      type="password"
-                      className={classnames("form-control", {
-                        invalid:
-                          errors.passwordConfirm || errors.passwordincorrect,
-                      })}
-                    />
-                  </div>
-                  <div
-                    style={{
-                      margin: 20,
-                    }}
-                    className="text-center pt-4"
-                  >
-                    <button type="submit" className="btn btn-outline-dark">
-                      Signup
-                    </button>
-                  </div>
-                  {/* <div className="text-center pt-2">
-                                            <a className="btn btn-link text-primary" href="#">Forgot Your Password?</a>
-                                        </div> */}
-                </form>
-              </div>
-            </div>
-          </center>
-        </main>
+// class Register extends Component {
+//   constructor() {
+//     super();
+//     this.state = {
+//       name: "",
+//       email: "",
+//       phone: "",
+//       password: "",
+//       passwordConfirm: "",
+//       errors: {},
+//     };
+//   }
+//   componentDidMount() {
+//     // If logged in and user navigates to Register page, should redirect them to
+//     // dashboard
+//     if (this.props.auth.isAuthenticated) {
+//       this.props.history.push("/");
+//     }
+//   }
+//   componentWillReceiveProps(nextProps) {
+//     if (nextProps.errors) {
+//       this.setState({ errors: nextProps.errors });
+//     }
+//   }
+//   onChange = (e) => {
+//     this.setState({
+//       [e.target.id]: e.target.value,
+//     });
+//   };
+//   onChangePhone = (e) => {
+//     this.setState({
+//       [e.target.id]: e.target.value.replace(/\D/,''),
+//     });
+//   };
+//   onSubmit = (e) => {
+//     // e.preventDefault();
+//     const newUser = {
+//       name: this.state.name,
+//       email: this.state.email,
+//       phone: this.state.phone,
+//       password: this.state.password,
+//       passwordConfirm: this.state.passwordConfirm,
+//     };
+//     this.props.registerUser(newUser, this.props.history);
+//     // console.log(newUser);
+//   };
+//   render() {
+//     const { errors } = this.state;
+//     return (
+//       <div className="App-login">
+//         <main>
+//           <center>
+//             <div className="container">
+//               <div
+//                 className="z-depth-1  lighten-4"
+//                 style={{
+//                   display: "inline-block",
+//                   padding: "32px 48px 0px 48px",
+//                   border: "1px solid #EEE",
+//                 }}
+//               >
+//                 <img
+//                   src={Logo}
+//                   className="responsive-img"
+//                   style={{
+//                     width: "150px",
+//                     margin: 20,
+//                   }}
+//                   alt="..."
+//                 />
+//                 <form onSubmit={this.onSubmit}>
+//                   <div className="form-group text-left">
+//                     <label>Full Name</label>
+//                     <input
+//                       className="form-control"
+//                       onChange={this.onChange}
+//                       value={this.state.name}
+//                       error={errors.name}
+//                       placeholder=" Full Name"
+//                       id="name"
+//                       size="25"
+//                       type="name"
+//                     />
+//                   </div>
+//                   <div className="form-group text-left">
+//                   <label>Email Address</label>
+//                     <input
+//                       onChange={this.onChange}
+//                       value={this.state.email}
+//                       error={errors.email}
+//                       placeholder=" Email Address"
+//                       id="email"
+//                       size="25"
+//                       type="email"
+//                       className={classnames("form-control", {
+//                         invalid: errors.email || errors.emailnotfound,
+//                       })}
+//                     />
+//                   </div>
+//                   <div className="form-group text-left">
+//                   <label>Phone Number</label>
+//                     <input
+//                       onChange={this.onChangePhone}
+//                       value={this.state.phone}
+//                       error={errors.phone}
+//                       placeholder=" Phone Number"
+//                       id="phone"
+//                       size="25"
+//                       type="text"
+//                       // pattern="[0-9]"
+//                       className={classnames("form-control", {})}
+//                     />
+//                   </div>
+//                   <div className="form-group text-left">
+//                   <label>Password</label>
+//                     <input
+//                       onChange={this.onChange}
+//                       value={this.state.password}
+//                       error={errors.password}
+//                       placeholder=" Password"
+//                       id="password"
+//                       size="25"
+//                       type="password"
+//                       className={classnames("form-control", {
+//                         invalid: errors.password || errors.passwordincorrect,
+//                       })}
+//                     />
+//                   </div>
+//                   <div className="form-group text-left">
+//                   <label>Confirm Password</label>
+//                     <input
+//                       onChange={this.onChange}
+//                       value={this.state.passwordConfirm}
+//                       error={errors.passwordConfirm}
+//                       placeholder=" Confirm Password"
+//                       id="passwordConfirm"
+//                       size="25"
+//                       type="password"
+//                       className={classnames("form-control", {
+//                         invalid:
+//                           errors.passwordConfirm || errors.passwordincorrect,
+//                       })}
+//                     />
+//                   </div>
+//                   <div
+//                     style={{
+//                       margin: 20,
+//                     }}
+//                     className="text-center pt-4"
+//                   >
+//                     <button type="submit" className="btn btn-outline-dark">
+//                       Signup
+//                     </button>
+//                   </div>
+//                   {/* <div className="text-center pt-2">
+//                                             <a className="btn btn-link text-primary" href="#">Forgot Your Password?</a>
+//                                         </div> */}
+//                 </form>
+//               </div>
+//             </div>
+//           </center>
+//         </main>
+//       </div>
+//     );
+//   }
+// }
+// Register.propTypes = {
+//   registerUser: PropTypes.func.isRequired,
+//   auth: PropTypes.object.isRequired,
+//   errors: PropTypes.object.isRequired,
+// };
+// const mapStateToProps = (state) => ({ auth: state.auth, errors: state.errors });
+// export default connect(mapStateToProps, { registerUser })(withRouter(Register));
+import React from "react";
+import "./SignIn.css";
+import { Button, Form, FormGroup, Label, Input } from "reactstrap";
+
+function SignUp() {
+  return (
+    <div className="d-md-flex flex-row">
+      <container className="SignIn header">
+        <h1 className="display-3 font-weight-bold">
+          Want to Buy Something Cheap?
+        </h1>
+        <h1>Take a Look Inside Here.</h1>
+        <p>
+          Lorem Ipsum is simply dummy text of the printing and typesetting
+          industry. Lorem Ipsum has been the industry's standard dummy text ever
+          since the 1500s, when an unknown printer took a galley of type and
+          scrambled it to make a type specimen book.
+        </p>
+      </container>
+      <div className="form">
+        <Form>
+          <FormGroup>
+          <h5 className="font-weight-bold">Full Name</h5>
+            <Input
+              className="wider"
+              type="name"
+              name="name"
+              placeholder="Your Name"
+            />
+          </FormGroup>
+          <FormGroup>
+          <h5 className="font-weight-bold">Email</h5>
+            <Input
+              className="wider"
+              type="email"
+              name="email"
+              placeholder="Password"
+            />
+            <h5 className="font-weight-bold">Password</h5>
+            <Input
+              className="wider"
+              type="email"
+              name="email"
+              placeholder="Email"
+            />
+          </FormGroup>
+          <FormGroup>
+          <h5 className="font-weight-bold">Confirm Password</h5>
+            <Input
+              className="wider"
+              type="password"
+              name="password"
+              placeholder="Password"
+            />
+          </FormGroup>
+          <FormGroup check>
+            <Label check>
+              <Input type="checkbox" />
+              Remember Me
+            </Label>
+          </FormGroup>
+          
+          <Button className="button center text-align-center box-wrapper">
+              Sign Up</Button>
+          <br /> 
+          <p className="password">
+              <a href="www.google.com">
+              Already have an account? Log here.<a href="" target="_blank"></a></a></p>
+        </Form>
       </div>
-    );
-  }
+    </div>
+  );
 }
-Register.propTypes = {
-  registerUser: PropTypes.func.isRequired,
-  auth: PropTypes.object.isRequired,
-  errors: PropTypes.object.isRequired,
-};
-const mapStateToProps = (state) => ({ auth: state.auth, errors: state.errors });
-export default connect(mapStateToProps, { registerUser })(withRouter(Register));
+
+export default SignUp;
